refactor: extract ui settings helpers
